Add comprehensive tests for GNFS math modules
