+++ tests/test_polynomial.py
@@ -1,5 +1,6 @@
 import pytest
-from gnfs.polynomial import Polynomial, select_polynomial
+from gnfs.polynomial.polynomial import Polynomial
+from gnfs.polynomial.selection import select_polynomial
 
 
 def test_polynomial_evaluate_and_degree():
@@ -29,3 +30,16 @@ def test_select_polynomial_higher_degree():
 def test_select_polynomial_invalid_degree():
     with pytest.raises(ValueError):
         select_polynomial(10, degree=0)
+
+
+def test_polynomial_constant_degree_zero():
+    poly = Polynomial((3,))
+    assert poly.degree() == 0
+    assert poly.evaluate(7) == 3
+
+
+def test_select_polynomial_root_property():
+    n = 10
+    poly = select_polynomial(n, degree=2)
+    m = round(n ** 0.5)
+    assert poly.evaluate(-m) % n == 0
